+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+# Latest updates
+
+From v1.6.6, You can check all of our changes from [Releae Page](https://github.com/Ikuyadeu/vscode-R/releases)
+
+## [1.6.6](https://github.com/Ikuyadeu/vscode-R/releases/tag/v1.6.6)
+
+Thank you for contributors
+* @krlmlr 
+  * Update vscode engine #586
+  * Satisfy markdownlint #587
+* @danielbasso
+  * Initial Workspace Viewer str() functionality #583
+
 ## 1.6.5
 
 * Add links to help pages in hover #578